requirements.testing.txt
pytest
pytest-cov
pytest-django
coverage
codecov
